requirements.txt
langchain
langchain-openai
langchain-community
azure-search-documents
python-dotenv
tiktoken
